requirements.txt
Django==3.1.2
requests==2.24.0
djangorestframework==3.12.1
psycopg2~=2.8.6
gunicorn==20.0.4
zeep==4.0.0
lxml==4.5.2
grpcio==1.32.0
pydantic==1.6.1
django-enumfield==2.0.2
PyCryptodome==3.9.9
cachetools==4.1.1
django-cors-headers
djangorestframework-api-key
django_extensions
django-sequences
stripe
haversine
pandas
rauth
django-redis
cryptography
pytz
django-admin-interface
xmltodict
https://github.com/chibisov/drf-extensions/archive/master.zip
django-jsonfield
django-allauth
django-rest-auth
django-rest-auth-bearer
django-storages[google]
google-cloud-storage==1.10.0
django-filter==2.4.0